refactor : 코드 중복 체크 로직 변경 및 시간복잡도 개선
