+++ Api/src/main/java/com/dingdong/api/community/service/CommunityService.java
@@ -8,7 +8,6 @@ import com.dingdong.api.community.dto.CommunityCodeDto;
 import com.dingdong.api.community.dto.CommunityDetailsDto;
 import com.dingdong.api.community.dto.CommunityIdCardsDto;
 import com.dingdong.api.community.dto.CommunityListDto;
-import com.dingdong.api.community.service.generator.GenerateCommunityInvitationCodeStrategy;
 import com.dingdong.api.community.service.generator.RandomCommunityCodeGeneratorStrategy;
 import com.dingdong.api.global.helper.UserHelper;
 import com.dingdong.api.idcard.dto.IdCardDetailsDto;
@@ -18,6 +17,7 @@ import com.dingdong.domain.common.util.SliceUtil;
 import com.dingdong.domain.domains.community.adaptor.CommunityAdaptor;
 import com.dingdong.domain.domains.community.domain.Community;
 import com.dingdong.domain.domains.community.domain.CommunityImage;
+import com.dingdong.domain.domains.community.domain.strategy.GenerateCommunityInvitationCodeStrategy;
 import com.dingdong.domain.domains.community.validator.CommunityValidator;
 import com.dingdong.domain.domains.idcard.adaptor.IdCardAdaptor;
 import com.dingdong.domain.domains.idcard.domain.entity.IdCard;
