+++ Api/src/main/java/com/dingdong/api/community/service/generator/RandomCommunityCodeGeneratorStrategy.java
@@ -1,23 +1,24 @@
 package com.dingdong.api.community.service.generator;
 
 
+import com.dingdong.domain.domains.community.domain.strategy.GenerateCommunityInvitationCodeStrategy;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import org.apache.commons.lang3.RandomStringUtils;
 
 public class RandomCommunityCodeGeneratorStrategy
         implements GenerateCommunityInvitationCodeStrategy {
 
-    private static final int MAX_RETRY = 10;
     private static final int CODE_LENGTH = 6;
 
-    public String generate(List<String> codes) {
-        String code = "";
-        for (int i = 0; i < MAX_RETRY; i++) {
+    public String generate(List<String> existingCodes) {
+        Set<String> codesSet = new HashSet<>(existingCodes);
+        String code;
+        do {
             code = RandomStringUtils.randomAlphanumeric(CODE_LENGTH);
-            if (!codes.contains(code)) {
-                break;
-            }
-        }
+        } while (codesSet.contains(code));
+
         return code;
     }
 }
